Implement changes from review
