+++ packages/lib/index.ts
@@ -18,13 +18,13 @@ export * from "./src/link";
 export * from "./src/logging";
 export * from "./src/logging_utils";
 export * from "./src/lua_tools";
+export * from "./src/permissions";
 export * from "./src/plugin";
 export * from "./src/plugin_loader";
 export * from "./src/prometheus";
 export * from "./src/schema";
 export * from "./src/shared_commands";
 export * from "./src/stream";
-export * from "./src/users";
 export * from "./src/zip_ops";
 export * from "./src/subscriptions";
