+++ packages/lib/src/permissions.ts
